cli: separate atsConsts from pTypes
