+++ packages/cli/src/portal/atsConsts.ts
@@ -1,0 +1,2 @@
+export const atsTypes = ["greenhouse", "lever"] as const;
+export type ATS = (typeof atsTypes)[number];
